\name{Truncated normal}
\alias{qtruncnorm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Truncated Normal Distribution
}
\description{
Quantile function for truncanted normal distribution with location and scale parameters}

\usage{
qtruncnorm(p, a = -Inf, b = Inf, mu = 0, sigma = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{vector of probabilities.}
  \item{a}{lower truncated value.}
  \item{b}{upper truncated value.}
  \item{mu}{location parameter can be given as vector.}
  \item{sigma}{Scale parameter can be given as vector.}
}
\details{
qtruncnorm gives the quantile function.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Israel Almodóvar-Rivera and Ranjan Maitra.
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
