\name{RFASTfMRI-package}
\alias{RFASTfMRI-package}
\alias{RFASTfMRI}
\docType{package}
\title{\packageTitle{RFASTfMRI}}
\description{\packageDescription{RFASTfMRI}}
\author{\packageAuthor{RFASTfMRI}}
\section{Maintainer}{\packageMaintainer{RFASTfMRI}}
\references{Almodovar-Rivera, I., & Maitra, R. (2019). FAST adaptive smoothing and thresholding for improved activation detection in low-signal fMRI. IEEE transactions on medical imaging, 38(12), 2821-2828.}
\keyword{ package }

