\name{RFASTfMRI-package}
\alias{RFASTfMRI-package}
\alias{RFASTfMRI}
\docType{package}
\title{\packageTitle{RFASTfMRI}}
\description{\packageDescription{RFASTfMRI}}
\author{\packageAuthor{RFASTfMRI}}
\section{Maintainer}{\packageMaintainer{RFASTfMRI}}
\references{Almodóvar-Rivera, I., & Maitra, R. (2017). FAST adaptive smoothing and thresholding for improved activation detection in low-signal fMRI. arXiv preprint arXiv:1702.00111.}
\keyword{ package }

